DDA-69: Fix grid create tool
